made modifications for the experimenter delegation model
In LoadGenerator there are three parameters: number of slices, initial level,
and branching factor. Every experimenter is given access to every slice, but at
a different length on a delegation chain. INITIAL_LEVEL configures how many
experimenters are one delegation away from having access. BRANCHING_FACTOR
tunes how many experimenters are at the next level. 2 for twice as many, 0.5
for half as many, etc. The number of levels is determined by how many
experimenters there are (set in IdentityManager). Levels are created until all
experimenters are exausted. This means the last level may not follow the
branching factor if the number of experimenters is insufficient.