changed load generation for GENI ORCA benchmark scenario.
The first time load generator is run, it will generate 1024 slices and
delegate them to all experimenters. The list of slices is saved to a
local file. Then a load generation loops, running access on random
experimenter/slice pairs indefinitely.

Subsequent runs will read in the list of slices to avoid the lengthy
delegation process.